Relocate services tests to test_services
