+++ tests/test_services.py
@@ -8,6 +8,7 @@ import logging
 import socket
 import threading
 import time
+import os
 import unittest
 from threading import Event
 
@@ -23,7 +24,7 @@ from zeroconf.services import (
     ServiceStateChange,
 )
 
-from . import _inject_response
+from . import has_working_ipv6, _clear_cache, _inject_response
 
 
 log = logging.getLogger('zeroconf')
@@ -535,6 +536,414 @@ class TestServiceBrowserMultipleTypes(unittest.TestCase):
             assert len(zeroconf.listeners) == 0
             zeroconf.remove_all_service_listeners()
             zeroconf.close()
+
+
+class ListenerTest(unittest.TestCase):
+    def test_integration_with_listener_class(self):
+
+        service_added = Event()
+        service_removed = Event()
+        service_updated = Event()
+        service_updated2 = Event()
+
+        subtype_name = "My special Subtype"
+        type_ = "_http._tcp.local."
+        subtype = subtype_name + "._sub." + type_
+        name = "UPPERxxxyyyæøå"
+        registration_name = "%s.%s" % (name, subtype)
+
+        class MyListener(r.ServiceListener):
+            def add_service(self, zeroconf, type, name):
+                zeroconf.get_service_info(type, name)
+                service_added.set()
+
+            def remove_service(self, zeroconf, type, name):
+                service_removed.set()
+
+            def update_service(self, zeroconf, type, name):
+                service_updated2.set()
+
+        class MySubListener(r.ServiceListener):
+            def add_service(self, zeroconf, type, name):
+                pass
+
+            def remove_service(self, zeroconf, type, name):
+                pass
+
+            def update_service(self, zeroconf, type, name):
+                service_updated.set()
+
+        listener = MyListener()
+        zeroconf_browser = Zeroconf(interfaces=['127.0.0.1'])
+        zeroconf_browser.add_service_listener(subtype, listener)
+
+        properties = dict(
+            prop_none=None,
+            prop_string=b'a_prop',
+            prop_float=1.0,
+            prop_blank=b'a blanked string',
+            prop_true=1,
+            prop_false=0,
+        )
+
+        zeroconf_registrar = Zeroconf(interfaces=['127.0.0.1'])
+        desc = {'path': '/~paulsm/'}  # type: Dict
+        desc.update(properties)
+        addresses = [socket.inet_aton("10.0.1.2")]
+        if has_working_ipv6() and not os.environ.get('SKIP_IPV6'):
+            addresses.append(socket.inet_pton(socket.AF_INET6, "6001:db8::1"))
+            addresses.append(socket.inet_pton(socket.AF_INET6, "2001:db8::1"))
+        info_service = ServiceInfo(
+            subtype, registration_name, port=80, properties=desc, server="ash-2.local.", addresses=addresses
+        )
+        zeroconf_registrar.register_service(info_service)
+
+        try:
+            service_added.wait(1)
+            assert service_added.is_set()
+
+            # short pause to allow multicast timers to expire
+            time.sleep(3)
+
+            # clear the answer cache to force query
+            _clear_cache(zeroconf_browser)
+
+            cached_info = ServiceInfo(type_, registration_name)
+            cached_info.load_from_cache(zeroconf_browser)
+            assert cached_info.properties == {}
+
+            # get service info without answer cache
+            info = zeroconf_browser.get_service_info(type_, registration_name)
+            assert info is not None
+            assert info.properties[b'prop_none'] is None
+            assert info.properties[b'prop_string'] == properties['prop_string']
+            assert info.properties[b'prop_float'] == b'1.0'
+            assert info.properties[b'prop_blank'] == properties['prop_blank']
+            assert info.properties[b'prop_true'] == b'1'
+            assert info.properties[b'prop_false'] == b'0'
+            assert info.addresses == addresses[:1]  # no V6 by default
+            assert info.addresses_by_version(r.IPVersion.All) == addresses
+
+            cached_info = ServiceInfo(type_, registration_name)
+            cached_info.load_from_cache(zeroconf_browser)
+            assert cached_info.properties is not None
+
+            # Populate the cache
+            zeroconf_browser.get_service_info(subtype, registration_name)
+
+            # get service info with only the cache
+            cached_info = ServiceInfo(subtype, registration_name)
+            cached_info.load_from_cache(zeroconf_browser)
+            assert cached_info.properties is not None
+            assert cached_info.properties[b'prop_float'] == b'1.0'
+
+            # get service info with only the cache with the lowercase name
+            cached_info = ServiceInfo(subtype, registration_name.lower())
+            cached_info.load_from_cache(zeroconf_browser)
+            # Ensure uppercase output is preserved
+            assert cached_info.name == registration_name
+            assert cached_info.key == registration_name.lower()
+            assert cached_info.properties is not None
+            assert cached_info.properties[b'prop_float'] == b'1.0'
+
+            info = zeroconf_browser.get_service_info(subtype, registration_name)
+            assert info is not None
+            assert info.properties is not None
+            assert info.properties[b'prop_none'] is None
+
+            cached_info = ServiceInfo(subtype, registration_name.lower())
+            cached_info.load_from_cache(zeroconf_browser)
+            assert cached_info.properties is not None
+            assert cached_info.properties[b'prop_none'] is None
+
+            # test TXT record update
+            sublistener = MySubListener()
+            zeroconf_browser.add_service_listener(registration_name, sublistener)
+            properties['prop_blank'] = b'an updated string'
+            desc.update(properties)
+            info_service = ServiceInfo(
+                subtype,
+                registration_name,
+                80,
+                0,
+                0,
+                desc,
+                "ash-2.local.",
+                addresses=[socket.inet_aton("10.0.1.2")],
+            )
+            zeroconf_registrar.update_service(info_service)
+            service_updated.wait(1)
+            assert service_updated.is_set()
+
+            info = zeroconf_browser.get_service_info(type_, registration_name)
+            assert info is not None
+            assert info.properties[b'prop_blank'] == properties['prop_blank']
+
+            cached_info = ServiceInfo(subtype, registration_name)
+            cached_info.load_from_cache(zeroconf_browser)
+            assert cached_info.properties is not None
+            assert cached_info.properties[b'prop_blank'] == properties['prop_blank']
+
+            zeroconf_registrar.unregister_service(info_service)
+            service_removed.wait(1)
+            assert service_removed.is_set()
+
+        finally:
+            zeroconf_registrar.close()
+            zeroconf_browser.remove_service_listener(listener)
+            zeroconf_browser.close()
+
+
+class TestServiceBrowser(unittest.TestCase):
+    def test_update_record(self):
+        enable_ipv6 = has_working_ipv6() and not os.environ.get('SKIP_IPV6')
+
+        service_name = 'name._type._tcp.local.'
+        service_type = '_type._tcp.local.'
+        service_server = 'ash-1.local.'
+        service_text = b'path=/~matt1/'
+        service_address = '10.0.1.2'
+        service_v6_address = "2001:db8::1"
+        service_v6_second_address = "6001:db8::1"
+
+        service_added_count = 0
+        service_removed_count = 0
+        service_updated_count = 0
+        service_add_event = Event()
+        service_removed_event = Event()
+        service_updated_event = Event()
+
+        class MyServiceListener(r.ServiceListener):
+            def add_service(self, zc, type_, name) -> None:
+                nonlocal service_added_count
+                service_added_count += 1
+                service_add_event.set()
+
+            def remove_service(self, zc, type_, name) -> None:
+                nonlocal service_removed_count
+                service_removed_count += 1
+                service_removed_event.set()
+
+            def update_service(self, zc, type_, name) -> None:
+                nonlocal service_updated_count
+                service_updated_count += 1
+                service_info = zc.get_service_info(type_, name)
+                assert socket.inet_aton(service_address) in service_info.addresses
+                if enable_ipv6:
+                    assert socket.inet_pton(
+                        socket.AF_INET6, service_v6_address
+                    ) in service_info.addresses_by_version(r.IPVersion.V6Only)
+                    assert socket.inet_pton(
+                        socket.AF_INET6, service_v6_second_address
+                    ) in service_info.addresses_by_version(r.IPVersion.V6Only)
+                assert service_info.text == service_text
+                assert service_info.server == service_server
+                service_updated_event.set()
+
+        def mock_incoming_msg(service_state_change: r.ServiceStateChange) -> r.DNSIncoming:
+
+            generated = r.DNSOutgoing(const._FLAGS_QR_RESPONSE)
+            assert generated.is_response() is True
+
+            if service_state_change == r.ServiceStateChange.Removed:
+                ttl = 0
+            else:
+                ttl = 120
+
+            generated.add_answer_at_time(
+                r.DNSText(
+                    service_name, const._TYPE_TXT, const._CLASS_IN | const._CLASS_UNIQUE, ttl, service_text
+                ),
+                0,
+            )
+
+            generated.add_answer_at_time(
+                r.DNSService(
+                    service_name,
+                    const._TYPE_SRV,
+                    const._CLASS_IN | const._CLASS_UNIQUE,
+                    ttl,
+                    0,
+                    0,
+                    80,
+                    service_server,
+                ),
+                0,
+            )
+
+            # Send the IPv6 address first since we previously
+            # had a bug where the IPv4 would be missing if the
+            # IPv6 was seen first
+            if enable_ipv6:
+                generated.add_answer_at_time(
+                    r.DNSAddress(
+                        service_server,
+                        const._TYPE_AAAA,
+                        const._CLASS_IN | const._CLASS_UNIQUE,
+                        ttl,
+                        socket.inet_pton(socket.AF_INET6, service_v6_address),
+                    ),
+                    0,
+                )
+                generated.add_answer_at_time(
+                    r.DNSAddress(
+                        service_server,
+                        const._TYPE_AAAA,
+                        const._CLASS_IN | const._CLASS_UNIQUE,
+                        ttl,
+                        socket.inet_pton(socket.AF_INET6, service_v6_second_address),
+                    ),
+                    0,
+                )
+            generated.add_answer_at_time(
+                r.DNSAddress(
+                    service_server,
+                    const._TYPE_A,
+                    const._CLASS_IN | const._CLASS_UNIQUE,
+                    ttl,
+                    socket.inet_aton(service_address),
+                ),
+                0,
+            )
+
+            generated.add_answer_at_time(
+                r.DNSPointer(service_type, const._TYPE_PTR, const._CLASS_IN, ttl, service_name), 0
+            )
+
+            return r.DNSIncoming(generated.packets()[0])
+
+        zeroconf = r.Zeroconf(interfaces=['127.0.0.1'])
+        service_browser = r.ServiceBrowser(zeroconf, service_type, listener=MyServiceListener())
+
+        try:
+            wait_time = 3
+
+            # service added
+            _inject_response(zeroconf, mock_incoming_msg(r.ServiceStateChange.Added))
+            service_add_event.wait(wait_time)
+            assert service_added_count == 1
+            assert service_updated_count == 0
+            assert service_removed_count == 0
+
+            # service SRV updated
+            service_updated_event.clear()
+            service_server = 'ash-2.local.'
+            _inject_response(zeroconf, mock_incoming_msg(r.ServiceStateChange.Updated))
+            service_updated_event.wait(wait_time)
+            assert service_added_count == 1
+            assert service_updated_count == 1
+            assert service_removed_count == 0
+
+            # service TXT updated
+            service_updated_event.clear()
+            service_text = b'path=/~matt2/'
+            _inject_response(zeroconf, mock_incoming_msg(r.ServiceStateChange.Updated))
+            service_updated_event.wait(wait_time)
+            assert service_added_count == 1
+            assert service_updated_count == 2
+            assert service_removed_count == 0
+
+            # service TXT updated - duplicate update should not trigger another service_updated
+            service_updated_event.clear()
+            service_text = b'path=/~matt2/'
+            _inject_response(zeroconf, mock_incoming_msg(r.ServiceStateChange.Updated))
+            service_updated_event.wait(wait_time)
+            assert service_added_count == 1
+            assert service_updated_count == 2
+            assert service_removed_count == 0
+
+            # service A updated
+            service_updated_event.clear()
+            service_address = '10.0.1.3'
+            _inject_response(zeroconf, mock_incoming_msg(r.ServiceStateChange.Updated))
+            service_updated_event.wait(wait_time)
+            assert service_added_count == 1
+            assert service_updated_count == 3
+            assert service_removed_count == 0
+
+            # service all updated
+            service_updated_event.clear()
+            service_server = 'ash-3.local.'
+            service_text = b'path=/~matt3/'
+            service_address = '10.0.1.3'
+            _inject_response(zeroconf, mock_incoming_msg(r.ServiceStateChange.Updated))
+            service_updated_event.wait(wait_time)
+            assert service_added_count == 1
+            assert service_updated_count == 4
+            assert service_removed_count == 0
+
+            # service removed
+            _inject_response(zeroconf, mock_incoming_msg(r.ServiceStateChange.Removed))
+            service_removed_event.wait(wait_time)
+            assert service_added_count == 1
+            assert service_updated_count == 4
+            assert service_removed_count == 1
+
+        finally:
+            assert len(zeroconf.listeners) == 1
+            service_browser.cancel()
+            assert len(zeroconf.listeners) == 0
+            zeroconf.remove_all_service_listeners()
+            zeroconf.close()
+
+
+def test_multiple_addresses():
+    type_ = "_http._tcp.local."
+    registration_name = "xxxyyy.%s" % type_
+    desc = {'path': '/~paulsm/'}
+    address_parsed = "10.0.1.2"
+    address = socket.inet_aton(address_parsed)
+
+    # New kwarg way
+    info = ServiceInfo(type_, registration_name, 80, 0, 0, desc, "ash-2.local.", addresses=[address, address])
+
+    assert info.addresses == [address, address]
+
+    info = ServiceInfo(
+        type_,
+        registration_name,
+        80,
+        0,
+        0,
+        desc,
+        "ash-2.local.",
+        parsed_addresses=[address_parsed, address_parsed],
+    )
+    assert info.addresses == [address, address]
+
+    if has_working_ipv6() and not os.environ.get('SKIP_IPV6'):
+        address_v6_parsed = "2001:db8::1"
+        address_v6 = socket.inet_pton(socket.AF_INET6, address_v6_parsed)
+        infos = [
+            ServiceInfo(
+                type_,
+                registration_name,
+                80,
+                0,
+                0,
+                desc,
+                "ash-2.local.",
+                addresses=[address, address_v6],
+            ),
+            ServiceInfo(
+                type_,
+                registration_name,
+                80,
+                0,
+                0,
+                desc,
+                "ash-2.local.",
+                parsed_addresses=[address_parsed, address_v6_parsed],
+            ),
+        ]
+        for info in infos:
+            assert info.addresses == [address]
+            assert info.addresses_by_version(r.IPVersion.All) == [address, address_v6]
+            assert info.addresses_by_version(r.IPVersion.V4Only) == [address]
+            assert info.addresses_by_version(r.IPVersion.V6Only) == [address_v6]
+            assert info.parsed_addresses() == [address_parsed, address_v6_parsed]
+            assert info.parsed_addresses(r.IPVersion.V4Only) == [address_parsed]
+            assert info.parsed_addresses(r.IPVersion.V6Only) == [address_v6_parsed]
 
 
 def test_backoff():
